refactor: rename hooks -> hooksStore (#4700)
