+++ apps/cowswap-frontend/src/modules/hooksStore/hookRegistry.tsx
