+++ apps/cowswap-frontend/src/modules/hooksStore/hooks/useRemoveHook.ts
