+++ apps/cowswap-frontend/src/modules/hooksStore/state/hookDetailsAtom.ts
